+++ test/segment_manager_test.cpp
@@ -212,7 +212,7 @@ bool test_segment_manager()
       seg_mgr->destroy_ptr(int_array);
       BOOST_TRY{  seg_mgr->template construct<int>(anonymous_instance)[ShmSizeSize]();  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct<int>(anonymous_instance, std::nothrow)[ShmSizeSize]())
-      BOOST_TRY{  seg_mgr->template construct_it<int>(anonymous_instance)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
+      BOOST_TRY{  seg_mgr->template construct_it<long int>(anonymous_instance)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct_it<int>(anonymous_instance, std::nothrow)[ShmSizeSize](&int_array_values[0]))
          return false;
       if(seg_mgr->get_free_memory() != free_mem_before)
@@ -297,7 +297,7 @@ bool test_segment_manager()
       }
       BOOST_TRY{  seg_mgr->template construct<unsigned int>(object1_name)[ShmSizeSize]();  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct<int>(object2_name, std::nothrow)[ShmSizeSize]())
-      BOOST_TRY{  seg_mgr->template construct_it<unsigned int>(object1_name)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
+      BOOST_TRY{  seg_mgr->template construct_it<int>(object1_name)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct_it<int>(object2_name, std::nothrow)[ShmSizeSize](&int_array_values[0]))
          return false;
       seg_mgr->shrink_to_fit_indexes();
@@ -387,7 +387,7 @@ bool test_segment_manager()
       }
       BOOST_TRY{  seg_mgr->template construct<unsigned int>(unique_instance)[ShmSizeSize]();  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct<int>(unique_instance, std::nothrow)[ShmSizeSize]())
-      BOOST_TRY{  seg_mgr->template construct_it<unsigned int>(unique_instance)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
+      BOOST_TRY{  seg_mgr->template construct_it<long int>(unique_instance)[ShmSizeSize](&int_array_values[0]);  }BOOST_CATCH(interprocess_exception&){} BOOST_CATCH_END
       if(seg_mgr->template construct_it<int>(unique_instance, std::nothrow)[ShmSizeSize](&int_array_values[0]))
          return false;
       seg_mgr->shrink_to_fit_indexes();
